- convert direction to signed before multiplication for the case enums are   unsigned
git-svn-id: svn://svn.code.sf.net/p/gwyddion/code/trunk/gwyddion@3699 25697fe4-d644-49b8-9d30-be5465f2adf3

+++ libprocess/simplefft.c
@@ -48,7 +48,7 @@ gwy_fft_hum(GwyTransformDirection dir,
     gint m, l, i, j, is;
     gdouble imlt;
 
-    imlt = dir * G_PI;
+    imlt = (gint)dir * G_PI;
     fact = 1.0/sqrt(n);
     j = 1;
     for (i = 1; i <= n; i++) {
